+++ src/test/kotlin/uk/gov/justice/digital/hmpps/locationsinsideprison/resource/LocationResourceIntTest.kt
@@ -933,256 +933,6 @@ class LocationResourceIntTest : SqsIntegrationTestBase() {
     }
   }
 
-  @DisplayName("PUT /locations/{id}/convert-cell-to-non-res-cell")
-  @Nested
-  inner class ConvertCellToNonResCell {
-
-    var convertCellToNonResidentialLocationRequest = ConvertCellToNonResidentialLocationRequest(
-      convertedCellType = ConvertedCellType.OTHER,
-      otherConvertedCellType = "Taning room",
-    )
-
-    @Nested
-    inner class Security {
-      @Test
-      fun `access forbidden when no authority`() {
-        webTestClient.put().uri("/locations/${cell1.id}/convert-cell-to-non-res-cell")
-          .exchange()
-          .expectStatus().isUnauthorized
-      }
-
-      @Test
-      fun `access forbidden when no role`() {
-        webTestClient.put().uri("/locations/${cell1.id}/convert-cell-to-non-res-cell")
-          .headers(setAuthorisation(roles = listOf()))
-          .header("Content-Type", "application/json")
-          .bodyValue(jsonString(convertCellToNonResidentialLocationRequest))
-          .exchange()
-          .expectStatus().isForbidden
-      }
-
-      @Test
-      fun `access forbidden with wrong role`() {
-        webTestClient.put().uri("/locations/${cell1.id}/convert-cell-to-non-res-cell")
-          .headers(setAuthorisation(roles = listOf("ROLE_BANANAS")))
-          .header("Content-Type", "application/json")
-          .bodyValue(jsonString(convertCellToNonResidentialLocationRequest))
-          .exchange()
-          .expectStatus().isForbidden
-      }
-
-      @Test
-      fun `access forbidden with right role, wrong scope`() {
-        webTestClient.put().uri("/locations/${cell1.id}/convert-cell-to-non-res-cell")
-          .headers(setAuthorisation(roles = listOf("ROLE_BANANAS"), scopes = listOf("read")))
-          .header("Content-Type", "application/json")
-          .bodyValue(jsonString(convertCellToNonResidentialLocationRequest))
-          .exchange()
-          .expectStatus().isForbidden
-      }
-    }
-
-    @Nested
-    inner class Validation {
-      @Test
-      fun `access client error bad data`() {
-        webTestClient.put().uri("/locations/${cell1.id}/convert-cell-to-non-res-cell")
-          .headers(setAuthorisation(roles = listOf("ROLE_MAINTAIN_LOCATIONS"), scopes = listOf("write")))
-          .header("Content-Type", "application/json")
-          .bodyValue("""{"prisonId": ""}""")
-          .exchange()
-          .expectStatus().is4xxClientError
-      }
-
-      @Test
-      fun `cannot update convert to non residential cell as ID is not found`() {
-        webTestClient.put().uri("/locations/${java.util.UUID.randomUUID()}/convert-cell-to-non-res-cell")
-          .headers(setAuthorisation(roles = listOf("ROLE_MAINTAIN_LOCATIONS"), scopes = listOf("write")))
-          .header("Content-Type", "application/json")
-          .bodyValue(jsonString(convertCellToNonResidentialLocationRequest))
-          .exchange()
-          .expectStatus().isEqualTo(404)
-      }
-    }
-
-    @Nested
-    inner class HappyPath {
-
-      @Test
-      fun `can update convert cell to non res cell`() {
-        prisonerSearchMockServer.stubSearchByLocations(cell1.prisonId, listOf(cell1.getPathHierarchy(), cell1.getPathHierarchy()), false)
-        val result = webTestClient.put().uri("/locations/${cell1.id}/convert-cell-to-non-res-cell")
-          .headers(setAuthorisation(roles = listOf("ROLE_MAINTAIN_LOCATIONS"), scopes = listOf("write")))
-          .header("Content-Type", "application/json")
-          .bodyValue(convertCellToNonResidentialLocationRequest)
-          .exchange()
-          .expectStatus().isOk
-          .expectBody(LocationTest::class.java)
-          .returnResult().responseBody!!
-
-        assertThat(result.findByPathHierarchy("Z-1-001")!!.convertedCellType == ConvertedCellType.OTHER)
-
-        getDomainEvents(3).let {
-          assertThat(it.map { message -> message.eventType to message.additionalInformation?.key }).containsExactlyInAnyOrder(
-            "location.inside.prison.amended" to "MDI-Z-1-001",
-            "location.inside.prison.amended" to "MDI-Z-1",
-            "location.inside.prison.amended" to "MDI-Z",
-          )
-        }
-      }
-    }
-  }
-
-  @DisplayName("PUT /locations/{id}/convert-to-cell")
-  @Nested
-  inner class ConvertToCellTest {
-    var convertToCellRequest = ConvertToCellRequest(
-      accommodationType = AccommodationType.NORMAL_ACCOMMODATION,
-      specialistCellType = SpecialistCellType.ACCESSIBLE_CELL,
-      maxCapacity = 2,
-      workingCapacity = 2,
-    )
-
-    var convertToCellRequestNotValidMaxCapacity = ConvertToCellRequest(
-      accommodationType = AccommodationType.CARE_AND_SEPARATION,
-      specialistCellType = SpecialistCellType.ACCESSIBLE_CELL,
-      maxCapacity = -1,
-      workingCapacity = 2,
-      usedForTypes = listOf(UsedForType.STANDARD_ACCOMMODATION, UsedForType.PERSONALITY_DISORDER),
-    )
-
-    var convertToCellRequestNotValidWorkingCapacity = ConvertToCellRequest(
-      accommodationType = AccommodationType.CARE_AND_SEPARATION,
-      specialistCellType = SpecialistCellType.ACCESSIBLE_CELL,
-      maxCapacity = 1,
-      workingCapacity = -1,
-      usedForTypes = listOf(UsedForType.STANDARD_ACCOMMODATION, UsedForType.PERSONALITY_DISORDER),
-    )
-
-    @Nested
-    inner class Security {
-      @Test
-      fun `access forbidden when no authority`() {
-        webTestClient.put().uri("/locations/${cell1.id}/convert-to-cell")
-          .exchange()
-          .expectStatus().isUnauthorized
-      }
-
-      @Test
-      fun `access forbidden when no role`() {
-        webTestClient.put().uri("/locations/${cell1.id}/convert-to-cell")
-          .headers(setAuthorisation(roles = listOf()))
-          .header("Content-Type", "application/json")
-          .bodyValue(jsonString(convertToCellRequest))
-          .exchange()
-          .expectStatus().isForbidden
-      }
-
-      @Test
-      fun `access forbidden with wrong role`() {
-        webTestClient.put().uri("/locations/${cell1.id}/convert-to-cell")
-          .headers(setAuthorisation(roles = listOf("ROLE_BANANAS")))
-          .header("Content-Type", "application/json")
-          .bodyValue(jsonString(convertToCellRequest))
-          .exchange()
-          .expectStatus().isForbidden
-      }
-
-      @Test
-      fun `access forbidden with right role, wrong scope`() {
-        webTestClient.put().uri("/locations/${cell1.id}/convert-to-cell")
-          .headers(setAuthorisation(roles = listOf("ROLE_BANANAS"), scopes = listOf("read")))
-          .header("Content-Type", "application/json")
-          .bodyValue(jsonString(convertToCellRequest))
-          .exchange()
-          .expectStatus().isForbidden
-      }
-    }
-
-    @Nested
-    inner class Validation {
-      @Test
-      fun `access client error bad data`() {
-        webTestClient.put().uri("/locations/${cell1.id}/convert-to-cell")
-          .headers(setAuthorisation(roles = listOf("ROLE_MAINTAIN_LOCATIONS"), scopes = listOf("write")))
-          .header("Content-Type", "application/json")
-          .bodyValue("""{"prisonId": ""}""")
-          .exchange()
-          .expectStatus().is4xxClientError
-      }
-
-      @Test
-      fun `cannot update convert to cell as Location ID is not found`() {
-        webTestClient.put().uri("/locations/${java.util.UUID.randomUUID()}/convert-to-cell")
-          .headers(setAuthorisation(roles = listOf("ROLE_MAINTAIN_LOCATIONS"), scopes = listOf("write")))
-          .header("Content-Type", "application/json")
-          .bodyValue(jsonString(convertToCellRequest))
-          .exchange()
-          .expectStatus().isEqualTo(404)
-      }
-
-      @Test
-      fun `cannot update convert to cell as request has invalid Max Capacity `() {
-        cell1.convertToNonResidentialCell(convertedCellType = ConvertedCellType.OTHER, userOrSystemInContext = "Aleman", clock = clock)
-        repository.save(cell1)
-        // request has not valid data
-        webTestClient.put().uri("/locations/${cell1.id}/convert-to-cell")
-          .headers(setAuthorisation(roles = listOf("ROLE_MAINTAIN_LOCATIONS"), scopes = listOf("write")))
-          .header("Content-Type", "application/json")
-          .bodyValue(convertToCellRequestNotValidMaxCapacity)
-          .exchange()
-          .expectStatus().isEqualTo(400)
-      }
-
-      @Test
-      fun `cannot update convert to cell as request has invalid Working Capacity `() { // request has not valid data
-        cell1.convertToNonResidentialCell(convertedCellType = ConvertedCellType.OTHER, userOrSystemInContext = "Aleman", clock = clock)
-        repository.save(cell1)
-
-        webTestClient.put().uri("/locations/${cell1.id}/convert-to-cell")
-          .headers(setAuthorisation(roles = listOf("ROLE_MAINTAIN_LOCATIONS"), scopes = listOf("write")))
-          .header("Content-Type", "application/json")
-          .bodyValue(convertToCellRequestNotValidWorkingCapacity)
-          .exchange()
-          .expectStatus().isEqualTo(400)
-      }
-    }
-
-    @Nested
-    inner class HappyPath {
-
-      // TODO MAP-1438 extra couple of tests need to be added for "Healthcare inpatients" & "Care and separation"
-      @Test
-      fun `can convert non-res cell to res cell`() {
-        cell1.convertToNonResidentialCell(convertedCellType = ConvertedCellType.OTHER, userOrSystemInContext = "Aleman", clock = clock)
-        repository.save(cell1)
-
-        val result = webTestClient.put().uri("/locations/${cell1.id}/convert-to-cell")
-          .headers(setAuthorisation(roles = listOf("ROLE_MAINTAIN_LOCATIONS"), scopes = listOf("write")))
-          .header("Content-Type", "application/json")
-          .bodyValue(convertToCellRequest)
-          .exchange()
-          .expectStatus().isOk
-          .expectBody(LocationTest::class.java)
-          .returnResult().responseBody!!
-
-        val cellZ1001 = result.findByPathHierarchy("Z-1-001")
-        assertThat(cellZ1001?.capacity?.maxCapacity).isEqualTo(2)
-        assertThat(cellZ1001?.capacity?.workingCapacity).isEqualTo(2)
-        assertThat(cellZ1001?.specialistCellTypes?.get(0)).isEqualTo(SpecialistCellType.ACCESSIBLE_CELL)
-        assertThat(cellZ1001?.convertedCellType).isNotEqualTo("OTHER")
-
-        getDomainEvents(3).let {
-          assertThat(it.map { message -> message.eventType to message.additionalInformation?.key }).containsExactlyInAnyOrder(
-            "location.inside.prison.amended" to "MDI-Z-1-001",
-            "location.inside.prison.amended" to "MDI-Z-1",
-            "location.inside.prison.amended" to "MDI-Z",
-          )
-        }
-      }
-    }
-  }
-
   @DisplayName("POST /locations/create-wing")
   @Nested
   inner class CreateWingTest {
